Add an improved version of FixAimingSpeed, needed for FPS above 60
