+++ dllmain/60fpsFixes.cpp
@@ -3,8 +3,11 @@
 #include "dllmain.h"
 #include "Settings.h"
 #include "ConsoleWnd.h"
+#include "KeyboardMouseTweaks.h"
 
 static uint32_t* ptrGameFrameRate;
+
+uintptr_t* ptrAimSpeedFldAddr;
 
 int intCurrentFrameRate()
 {
@@ -76,6 +79,32 @@ void Init_60fpsFixes()
 				_asm {fmul vanillaMulti}
 		}
 	}; injector::MakeInline<AmmoBoxSpeed>(pattern.count(2).get(1).get<uint32_t>(0), pattern.count(2).get(1).get<uint32_t>(6));
+
+	// Aiming speed
+	pattern = hook::pattern("DD 05 ? ? ? ? DC F9 DD 05 ? ? ? ? DC FA D9 01 DE CB D9 45");
+	ptrAimSpeedFldAddr = *pattern.count(1).get(0).get<uint32_t*>(2);
+	struct AimSpeed
+	{
+		void operator()(injector::reg_pack& regs)
+		{
+			double aim_spd_val = *(double*)ptrAimSpeedFldAddr;
+
+			if (cfg.bFixAimingSpeed)
+				aim_spd_val = static_cast<double>(6000) / intCurrentFrameRate();
+
+			// Apply custom controller sensitivity
+			if ((GetLastUsedDevice() == LastDevice::XinputController) || (GetLastUsedDevice() == LastDevice::DinputController))
+				aim_spd_val /= cfg.fControllerSensitivity;
+
+			_asm
+			{
+				fld aim_spd_val
+			}
+		}
+	}; injector::MakeInline<AimSpeed>(pattern.count(1).get(0).get<uint32_t>(0), pattern.count(1).get(0).get<uint32_t>(6));
+
+	pattern = hook::pattern("DD 05 ? ? ? ? DC F9 DD 05 ? ? ? ? DC FA D9 01");
+	injector::MakeInline<AimSpeed>(pattern.count(1).get(0).get<uint32_t>(0), pattern.count(1).get(0).get<uint32_t>(6));
 
 	// Fix games framelimiter to improve support for framerates greater than 60
 	pattern = hook::pattern("00 00 00 20 11 11 91 3F"); // default is 0.1666666..., luckily only one instance of it in the game EXE
