--- dllmain/ControllerTweaks.h
@@ -1,3 +1,0 @@
-#pragma once
-
-void Init_ControllerTweaks();
