+++ dllmain/dllmain.vcxproj.filters
@@ -88,9 +88,6 @@
     <ClCompile Include="ExceptionHandler.cpp">
       <Filter>dllmain</Filter>
     </ClCompile>
-    <ClCompile Include="ControllerTweaks.cpp">
-      <Filter>dllmain</Filter>
-    </ClCompile>
     <ClCompile Include="..\external\imgui\imgui_stdlib.cpp">
       <Filter>includes\external\ImGui</Filter>
     </ClCompile>
@@ -311,9 +308,6 @@
     </ClInclude>
     <ClInclude Include="..\Wrappers\x3daudio1_7.h">
       <Filter>Wrappers</Filter>
-    </ClInclude>
-    <ClInclude Include="ControllerTweaks.h">
-      <Filter>dllmain</Filter>
     </ClInclude>
     <ClInclude Include="..\external\imgui\imgui_stdlib.h">
       <Filter>includes\external\ImGui</Filter>
