+++ settings/settings_string.h
@@ -128,13 +128,16 @@ QTE_key_2 = A
 FallbackToEnglishKeyIcons = true
 
 [FRAME RATE]
-; Fixes the speed of falling items in 60 FPS, making them not fall at double speed.
-FixFallingItemsSpeed = true
-
 ; When running in 60 FPS, some QTEs require extremely fast button presses to work. This gets even worse in Professional difficulty,
 ; making it seem almost impossible to survive the minecart and the statue bridge QTEs.
 ; This fix makes QTEs that involve rapid button presses much more forgiving.
 FixQTE = true
+
+; Fixes the speed of falling items in 60 FPS, making them not fall at double speed.
+FixFallingItemsSpeed = true
+
+; Fixes the movement speed when aiming weapons being slower on every frame rate above 30 FPS.
+FixAimingSpeed = true
 
 ; Fixes difference between 30/60FPS on physics applied to Ashley.
 FixAshleyBustPhysics = true
